# app/Models/Product.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Product extends Model
{
    protected $fillable = ['title', 'description', 'price', 'color', 'composition', 'images'];

    protected $casts = [
        'images' => 'array',  
    ];
}